split circ simulator into multiple files, add custom histogram parsing for compliance with Qiskits' format
