+++ cirq_simulator/__init__.py
@@ -1,0 +1,58 @@
+# Copyright 2023 QHAna plugin runner contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Optional
+
+from flask import Flask
+
+from qhana_plugin_runner.api.util import SecurityBlueprint
+from qhana_plugin_runner.util.plugins import plugin_identifier, QHAnaPluginBase
+
+_plugin_name = "cirq-simulator"
+__version__ = "v1.0.0"
+_identifier = plugin_identifier(_plugin_name, __version__)
+
+CIRQ_BLP = SecurityBlueprint(
+    _identifier,  # blueprint name
+    __name__,  # module import name!
+    description="Circuit executor exposing the cirq simulators as backend.",
+)
+
+
+class CirqSimulator(QHAnaPluginBase):
+    name = _plugin_name
+    version = __version__
+    description = (
+        "Allows execution of quantum circuits using a simulator packaged with cirq."
+    )
+    tags = ["circuit-executor", "qc-simulator", "cirq", "qasm", "qasm-2"]
+
+    def __init__(self, app: Optional[Flask]) -> None:
+        super().__init__(app)
+
+    def get_api_blueprint(self):
+        return CIRQ_BLP
+
+    def get_requirements(self) -> str:
+        return "cirq~=1.3.0.dev20230804185427"
+
+
+try:
+    # It is important to import the routes **after** COSTUME_LOADER_BLP and CostumeLoader are defined, because they are
+    # accessed as soon as the routes are imported.
+    from . import routes
+except ImportError:
+    # When running `poetry run flask install`, importing the routes will fail, because the dependencies are not
+    # installed yet.
+    pass
