+++ cirq_simulator/routes.py
@@ -1,0 +1,222 @@
+# Copyright 2023 QHAna plugin runner contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from collections import ChainMap
+from http import HTTPStatus
+from json import dumps
+from typing import Any, Dict, Mapping
+
+from celery.canvas import chain
+from flask import abort, redirect
+from flask.globals import request
+from flask.helpers import url_for
+from flask.templating import render_template
+from flask.views import MethodView
+from flask.wrappers import Response
+from marshmallow import EXCLUDE
+
+from qhana_plugin_runner.api.plugin_schemas import (
+    EntryPoint,
+    InputDataMetadata,
+    OutputDataMetadata,
+    PluginMetadata,
+    PluginMetadataSchema,
+    PluginType,
+)
+from qhana_plugin_runner.api.util import (
+    SecurityBlueprint,
+)
+from qhana_plugin_runner.db.models.tasks import ProcessingTask
+from qhana_plugin_runner.tasks import save_task_error, save_task_result
+from qhana_plugin_runner.util.plugins import plugin_identifier
+from . import CirqSimulator
+from .schemas import CirqSimulatorParametersSchema
+from .tasks import execute_circuit
+
+
+_plugin_name = "cirq-simulator"
+__version__ = "v1.0.0"
+_identifier = plugin_identifier(_plugin_name, __version__)
+
+CIRQ_BLP = SecurityBlueprint(
+    _identifier,  # blueprint name
+    __name__,  # module import name!
+    description="Circuit executor exposing the cirq simulators as backend.",
+)
+
+
+@CIRQ_BLP.route("/")
+class PluginsView(MethodView):
+    """Plugins collection resource."""
+
+    @CIRQ_BLP.response(HTTPStatus.OK, PluginMetadataSchema())
+    @CIRQ_BLP.require_jwt("jwt", optional=True)
+    def get(self):
+        """Endpoint returning the plugin metadata."""
+        plugin = CirqSimulator.instance
+        if plugin is None:
+            abort(HTTPStatus.INTERNAL_SERVER_ERROR)
+        return PluginMetadata(
+            title=plugin.name,
+            description=plugin.description,
+            name=plugin.name,
+            version=plugin.version,
+            type=PluginType.processing,
+            entry_point=EntryPoint(
+                href=url_for(f"{CIRQ_BLP.name}.ProcessView"),
+                ui_href=url_for(f"{CIRQ_BLP.name}.MicroFrontend"),
+                plugin_dependencies=[],
+                data_input=[
+                    InputDataMetadata(
+                        data_type="executable/circuit",
+                        content_type=["text/x-qasm"],
+                        required=True,
+                        parameter="circuit",
+                    ),
+                    InputDataMetadata(
+                        data_type="provenance/execution-options",
+                        content_type=[
+                            "text/csv",
+                            "application/json",
+                            "application/X-lines+json",
+                        ],
+                        required=False,
+                        parameter="executionOptions",
+                    ),
+                ],
+                data_output=[
+                    OutputDataMetadata(
+                        data_type="entity/vector",
+                        content_type=["application/json"],
+                        required=True,
+                        name="result-counts.json",
+                    ),
+                    OutputDataMetadata(
+                        data_type="entity/vector",
+                        content_type=["application/json"],
+                        required=False,
+                        name="result-statevector.json",
+                    ),
+                    OutputDataMetadata(
+                        data_type="provenance/trace",
+                        content_type=["application/json"],
+                        required=True,
+                        name="result-trace.json",
+                    ),
+                    OutputDataMetadata(
+                        data_type="provenance/execution-options",
+                        content_type=["application/json"],
+                        required=True,
+                        name="execution-options.json",
+                    ),
+                ],
+            ),
+            tags=CirqSimulator.instance.tags,
+        )
+
+
+################
+
+
+@CIRQ_BLP.route("/ui/")
+class MicroFrontend(MethodView):
+    """Micro frontend for the cirq simulators plugin."""
+
+    example_inputs: Dict[str, Any] = {
+        "shots": 1024,
+    }
+
+    @CIRQ_BLP.html_response(
+        HTTPStatus.OK, description="Micro frontend of the cirq simulators plugin."
+    )
+    @CIRQ_BLP.arguments(
+        CirqSimulatorParametersSchema(
+            partial=True, unknown=EXCLUDE, validate_errors_as_result=True
+        ),
+        location="query",
+        required=False,
+    )
+    @CIRQ_BLP.require_jwt("jwt", optional=True)
+    def get(self, errors):
+        """Return the micro frontend."""
+        values: ChainMap[str, Any] = ChainMap(request.args.to_dict(), self.example_inputs)
+        return self.render(values, errors, False)
+
+    @CIRQ_BLP.html_response(
+        HTTPStatus.OK, description="Micro frontend of the cirq simulators plugin."
+    )
+    @CIRQ_BLP.arguments(
+        CirqSimulatorParametersSchema(
+            partial=True, unknown=EXCLUDE, validate_errors_as_result=True
+        ),
+        location="form",
+        required=False,
+    )
+    @CIRQ_BLP.require_jwt("jwt", optional=True)
+    def post(self, errors):
+        """Return the micro frontend with prerendered inputs."""
+        values: ChainMap[str, Any] = ChainMap(request.form.to_dict(), self.example_inputs)
+        return self.render(values, errors, not errors)
+
+    def render(self, data: Mapping, errors: dict, valid: bool):
+        plugin = CirqSimulator.instance
+        if plugin is None:
+            abort(HTTPStatus.INTERNAL_SERVER_ERROR)
+        schema = CirqSimulatorParametersSchema()
+        return Response(
+            render_template(
+                "simple_template.html",
+                name=plugin.name,
+                version=plugin.version,
+                schema=schema,
+                valid=valid,
+                values=data,
+                errors=errors,
+                process=url_for(f"{CIRQ_BLP.name}.ProcessView"),
+                help_text="",
+                example_values=url_for(
+                    f"{CIRQ_BLP.name}.MicroFrontend", **self.example_inputs
+                ),
+            )
+        )
+
+
+@CIRQ_BLP.route("/process/")
+class ProcessView(MethodView):
+    """Start a long running processing task."""
+
+    @CIRQ_BLP.arguments(CirqSimulatorParametersSchema(unknown=EXCLUDE), location="form")
+    @CIRQ_BLP.response(HTTPStatus.FOUND)
+    @CIRQ_BLP.require_jwt("jwt", optional=True)
+    def post(self, arguments):
+        """Start the circuit execution task."""
+        db_task = ProcessingTask(
+            task_name=execute_circuit.name, parameters=dumps(arguments)
+        )
+        db_task.save(commit=True)
+
+        # all tasks need to know about db id to load the db entry
+        task: chain = execute_circuit.s(db_id=db_task.id) | save_task_result.s(
+            db_id=db_task.id
+        )
+        # save errors to db
+        task.link_error(save_task_error.s(db_id=db_task.id))
+        task.apply_async()
+
+        db_task.save(commit=True)
+
+        return redirect(
+            url_for("tasks-api.TaskView", task_id=str(db_task.id)), HTTPStatus.SEE_OTHER
+        )
