+++ cirq_simulator/schemas.py
@@ -1,0 +1,53 @@
+import marshmallow as ma
+from marshmallow.validate import Range
+
+from qhana_plugin_runner.api.util import (
+    FileUrl,
+    FrontendFormBaseSchema,
+)
+
+
+class CirqSimulatorParametersSchema(FrontendFormBaseSchema):
+    circuit = FileUrl(
+        required=True,
+        allow_none=False,
+        data_input_type="executable/circuit",
+        data_content_types="text/x-qasm",
+        metadata={
+            "label": "OpenQASM Circuit",
+            "description": "URL to a quantum circuit in the OpenQASM format.",
+            "input_type": "text",
+        },
+    )
+    executionOptions = FileUrl(
+        required=False,
+        allow_none=True,
+        load_missing=None,
+        data_input_type="provenance/execution-options",
+        data_content_types=["text/csv", "application/json", "application/X-lines+json"],
+        metadata={
+            "label": "Execution Options (optional)",
+            "description": "URL to a file containing execution options. (optional)",
+            "input_type": "text",
+        },
+    )
+    shots = ma.fields.Integer(
+        required=False,
+        allow_none=True,
+        load_default=None,
+        validate=Range(min=1, min_inclusive=True),
+        metadata={
+            "label": "Shots",
+            "description": "The number of shots to simulate. If execution options are specified they will override this setting!",
+            "input_type": "number",
+        },
+    )
+    statevector = ma.fields.Bool(
+        required=False,
+        allow_none=True,
+        load_default=False,
+        metadata={
+            "label": "Include Statevector",
+            "description": "Include a statevector result.",
+        },
+    )
