+++ cirq_simulator/tasks.py
@@ -1,0 +1,145 @@
+import mimetypes
+from json import dump, dumps, loads
+from tempfile import SpooledTemporaryFile
+from typing import Any, Dict, Optional, Union, cast
+from uuid import uuid4
+
+from celery.utils.log import get_task_logger
+
+from qhana_plugin_runner.celery import CELERY
+from qhana_plugin_runner.db.models.tasks import ProcessingTask
+from qhana_plugin_runner.plugin_utils.entity_marshalling import (
+    ensure_dict,
+    load_entities,
+)
+from qhana_plugin_runner.requests import open_url
+from qhana_plugin_runner.storage import STORE
+from . import CirqSimulator, _identifier
+from .utils import simulate_circuit
+
+TASK_LOGGER = get_task_logger(__name__)
+
+
+@CELERY.task(name=f"{CirqSimulator.instance.identifier}.demo_task", bind=True)
+def execute_circuit(self, db_id: int) -> str:
+    task_data: Optional[ProcessingTask] = ProcessingTask.get_by_id(id_=db_id)
+
+    if task_data is None:
+        msg = f"Could not load task data with id {db_id} to read parameters!"
+        TASK_LOGGER.error(msg)
+        raise KeyError(msg)
+
+    task_options: Dict[str, Union[str, int]] = loads(task_data.parameters or "{}")
+    circuit_url: str = cast(str, task_options["circuit"])
+
+    circuit_qasm: str
+    with open_url(circuit_url) as quasm_response:
+        circuit_qasm = quasm_response.text
+
+    execution_options_url = cast(
+        Optional[str], task_options.get("executionOptions", None)
+    )
+
+    execution_options: Dict[str, Any] = {
+        "shots": task_options.get("shots", 1),
+        "statevector": bool(task_options.get("statevector")),
+    }
+
+    if execution_options_url:
+        with open_url(execution_options_url) as execution_options_response:
+            try:
+                mimetype = execution_options_response.headers["Content-Type"]
+            except KeyError:
+                mimetype = mimetypes.MimeTypes().guess_type(url=execution_options_url)[0]
+            if mimetype is None:
+                msg = "Could not guess execution options mime type!"
+                TASK_LOGGER.error(msg)
+                raise ValueError(msg)
+            entities = ensure_dict(
+                load_entities(execution_options_response, mimetype=mimetype)
+            )
+            options = next(entities, {})
+            task_data.add_task_log_entry(
+                "loaded execution options: " + dumps(options), commit=True
+            )
+            execution_options.update(options)
+
+    if isinstance(execution_options["shots"], str):
+        execution_options["shots"] = int(execution_options["shots"])
+    if isinstance(execution_options["statevector"], str):
+        execution_options["statevector"] = execution_options["ststevector"] in (
+            "1",
+            "yes",
+            "Yes",
+            "YES",
+            "true",
+            "True",
+            "TRUE",
+        )
+
+    metadata, binary_histogram, state_vector = simulate_circuit(
+        circuit_qasm, execution_options
+    )
+
+    binary_histogram = {k: int(v) for k, v in binary_histogram.items()}
+
+    experiment_id = str(uuid4())
+
+    with SpooledTemporaryFile(mode="w") as output:
+        metadata["ID"] = experiment_id
+        dump(metadata, output)
+        STORE.persist_task_result(
+            db_id, output, "result-trace.json", "provenance/trace", "application/json"
+        )
+
+    with SpooledTemporaryFile(mode="w") as output:
+        binary_histogram["ID"] = experiment_id
+        dump(binary_histogram, output)
+        STORE.persist_task_result(
+            db_id, output, "result-counts.json", "entity/vector", "application/json"
+        )
+
+    # Finally, #if the conditions are ok, the `state_vector` is converted into a list,  and each element is transformed into a string, This can be used to convert complex numbers into the string format.
+
+    if state_vector is not None and any(state_vector):
+        state_vector_ent = {"ID": experiment_id}
+        str_vector = [str(x) for x in state_vector.tolist()]  #### tolist here ok!
+        dim = len(str_vector)
+        key_len = len(str(dim))
+        for i, v in enumerate(str_vector):
+            state_vector_ent[f"{i:0{key_len}}"] = repr(v)
+
+        with SpooledTemporaryFile(mode="w") as output:
+            dump(state_vector_ent, output)
+            STORE.persist_task_result(
+                db_id,
+                output,
+                "result-statevector.json",
+                "entity/vector",
+                "application/json",
+            )
+
+    extra_execution_options = {
+        "ID": experiment_id,
+        "executorPlugin": execution_options.get("executorPlugin", []) + [_identifier],
+        "shots": metadata.get("shots", execution_options["shots"]),
+        "simulator": metadata["qpuType"],
+        "qpuType": metadata["qpuType"],
+        "qpuVendor": metadata["qpuVendor"],
+        "qpuName": metadata["qpuName"],
+        "qpuVersion": metadata["qpuVersion"],
+    }
+
+    execution_options.update(extra_execution_options)
+
+    with SpooledTemporaryFile(mode="w") as output:
+        dump(execution_options, output)
+        STORE.persist_task_result(
+            db_id,
+            output,
+            "execution-options.json",
+            "provenance/execution-options",
+            "application/json",
+        )
+
+    return "Finished simulating circuit."
